+++ jobbergate-api/CHANGELOG.rst
@@ -8,6 +8,7 @@ Unreleased
 ----------
 - Job scripts were moved from a database column to files at S3.
 - Add email notification to Jobbergate.
+- More debug log messages were added to the API.
 
 3.2.0 -- 2022-06-24
 -------------------